Removed spacy_model and _model, removed relevant functionality from __main__
